+++ src/HttpCache/ETagValueNormalizer/NoopETagValueNormalizer.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\HttpMiddleware\HttpCache\ETagValueNormalizer;
+
+/**
+ * Returns ETag values unmodified. It can be used when ETag normalization is not required.
+ */
+final class NoopETagValueNormalizer implements ETagValueNormalizerInterface
+{
+    public function normalize(string $value): string
+    {
+        return $value;
+    }
+}